Bump version number to 3.0.0
In 4ac00c2 we replaced the use of Buffer.get with an array index. I'm
not sure if the Buffer array index has always existed, or if it was
added after Buffer.get...

... If it was added after Buffer.get, it's possible that 4ac00c2 is a
breaking a change. Because I'm unsure, I'm going to air on the side of
caution and bump the major version number.